backend/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6

# 视频处理
ffmpeg-python==0.2.0
opencv-python==4.9.0.80
scenedetect[opencv]>=0.6.0

# 音频处理 (ASR)
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# OCR
paddleocr==2.7.3
paddlepaddle>=2.6.2

# 对象检测
ultralytics==8.1.0  # YOLOv8

# AI API
anthropic==0.18.0
openai>=1.0.0  # 支持 OpenAI 兼容的第三方平台（如 aihubmix）

# 数据库
sqlalchemy==2.0.25

# 工具
python-dotenv==1.0.0
loguru==0.7.2
Pillow==10.2.0
numpy==1.26.3
